+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Download this extension from the [VSGallery](https://marketplace.visualstudio.co
 - [ ] Visual Studio 2019 support
 - [ ] Better support for new project format
 - [ ] Type objects for retrieving values
+- [x] Path of generator file appears in sql output (comment).
 
 Features that have a checkmark are complete and available for
 download in the [CI build](http://vsixgallery.com/author/danny%20quinn).